+++ src/index.ts
@@ -3,3 +3,4 @@ export * from './prop';
 export * from './Restgoose';
 export * from './middleware-composers';
 export { getModel } from './getModel';
+export { RestRequest } from './types';